+++ Library.Net.Amoeba/ConnectionManager.cs
@@ -959,7 +959,7 @@ namespace Library.Net.Amoeba
                 {
                     using (var stream = value.Export(bufferManager))
                     {
-                        ItemUtilities.Write(bufferStream, (byte)SerializeId.Node, stream);
+                        ItemUtilities.Write(bufferStream, (int)SerializeId.Node, stream);
                     }
                 }
 
@@ -1038,7 +1038,7 @@ namespace Library.Net.Amoeba
                 {
                     using (var stream = value.Export(bufferManager))
                     {
-                        ItemUtilities.Write(bufferStream, (byte)SerializeId.Key, stream);
+                        ItemUtilities.Write(bufferStream, (int)SerializeId.Key, stream);
                     }
                 }
 
@@ -1117,7 +1117,7 @@ namespace Library.Net.Amoeba
                 {
                     using (var stream = value.Export(bufferManager))
                     {
-                        ItemUtilities.Write(bufferStream, (byte)SerializeId.Key, stream);
+                        ItemUtilities.Write(bufferStream, (int)SerializeId.Key, stream);
                     }
                 }
 
@@ -1227,7 +1227,7 @@ namespace Library.Net.Amoeba
                 {
                     using (var stream = this.Key.Export(bufferManager))
                     {
-                        ItemUtilities.Write(bufferStream, (byte)SerializeId.Key, stream);
+                        ItemUtilities.Write(bufferStream, (int)SerializeId.Key, stream);
                     }
                 }
                 // Value
@@ -1317,7 +1317,7 @@ namespace Library.Net.Amoeba
                 // Signatures
                 foreach (var value in this.Signatures)
                 {
-                    ItemUtilities.Write(bufferStream, (byte)SerializeId.Signature, value);
+                    ItemUtilities.Write(bufferStream, (int)SerializeId.Signature, value);
                 }
 
                 bufferStream.Seek(0, SeekOrigin.Begin);
@@ -1395,7 +1395,7 @@ namespace Library.Net.Amoeba
                 {
                     using (var stream = value.Export(bufferManager))
                     {
-                        ItemUtilities.Write(bufferStream, (byte)SerializeId.Seed, stream);
+                        ItemUtilities.Write(bufferStream, (int)SerializeId.Seed, stream);
                     }
                 }
