+++ Library.Net.Outopos/Cache/Message/Items/MulticastMessage.cs
@@ -85,19 +85,19 @@ namespace Library.Net.Outopos
             {
                 using (var stream = this.Tag.Export(bufferManager))
                 {
-                    ItemUtilities.Write(bufferStream, (byte)SerializeId.Tag, stream);
+                    ItemUtilities.Write(bufferStream, (int)SerializeId.Tag, stream);
                 }
             }
             // CreationTime
             if (this.CreationTime != DateTime.MinValue)
             {
-                ItemUtilities.Write(bufferStream, (byte)SerializeId.CreationTime, this.CreationTime.ToUniversalTime().ToString("yyyy-MM-ddTHH:mm:ssZ", System.Globalization.DateTimeFormatInfo.InvariantInfo));
+                ItemUtilities.Write(bufferStream, (int)SerializeId.CreationTime, this.CreationTime.ToUniversalTime().ToString("yyyy-MM-ddTHH:mm:ssZ", System.Globalization.DateTimeFormatInfo.InvariantInfo));
             }
 
             // Comment
             if (this.Comment != null)
             {
-                ItemUtilities.Write(bufferStream, (byte)SerializeId.Comment, this.Comment);
+                ItemUtilities.Write(bufferStream, (int)SerializeId.Comment, this.Comment);
             }
 
             // Certificate
@@ -105,7 +105,7 @@ namespace Library.Net.Outopos
             {
                 using (var stream = this.Certificate.Export(bufferManager))
                 {
-                    ItemUtilities.Write(bufferStream, (byte)SerializeId.Certificate, stream);
+                    ItemUtilities.Write(bufferStream, (int)SerializeId.Certificate, stream);
                 }
             }
