+++ Library.Net.Outopos/Cache/Message/Items/UnicastMessage.cs
@@ -83,18 +83,18 @@ namespace Library.Net.Outopos
             // Signature
             if (this.Signature != null)
             {
-                ItemUtilities.Write(bufferStream, (byte)SerializeId.Signature, this.Signature);
+                ItemUtilities.Write(bufferStream, (int)SerializeId.Signature, this.Signature);
             }
             // CreationTime
             if (this.CreationTime != DateTime.MinValue)
             {
-                ItemUtilities.Write(bufferStream, (byte)SerializeId.CreationTime, this.CreationTime.ToUniversalTime().ToString("yyyy-MM-ddTHH:mm:ssZ", System.Globalization.DateTimeFormatInfo.InvariantInfo));
+                ItemUtilities.Write(bufferStream, (int)SerializeId.CreationTime, this.CreationTime.ToUniversalTime().ToString("yyyy-MM-ddTHH:mm:ssZ", System.Globalization.DateTimeFormatInfo.InvariantInfo));
             }
 
             // Comment
             if (this.Comment != null)
             {
-                ItemUtilities.Write(bufferStream, (byte)SerializeId.Comment, this.Comment);
+                ItemUtilities.Write(bufferStream, (int)SerializeId.Comment, this.Comment);
             }
 
             // Certificate
@@ -102,7 +102,7 @@ namespace Library.Net.Outopos
             {
                 using (var stream = this.Certificate.Export(bufferManager))
                 {
-                    ItemUtilities.Write(bufferStream, (byte)SerializeId.Certificate, stream);
+                    ItemUtilities.Write(bufferStream, (int)SerializeId.Certificate, stream);
                 }
             }
