+++ Library.Net.Outopos/Cache/Metadata/Items/MulticastMetadata.cs
@@ -90,13 +90,13 @@ namespace Library.Net.Outopos
             {
                 using (var stream = this.Tag.Export(bufferManager))
                 {
-                    ItemUtilities.Write(bufferStream, (byte)SerializeId.Tag, stream);
+                    ItemUtilities.Write(bufferStream, (int)SerializeId.Tag, stream);
                 }
             }
             // CreationTime
             if (this.CreationTime != DateTime.MinValue)
             {
-                ItemUtilities.Write(bufferStream, (byte)SerializeId.CreationTime, this.CreationTime.ToUniversalTime().ToString("yyyy-MM-ddTHH:mm:ssZ", System.Globalization.DateTimeFormatInfo.InvariantInfo));
+                ItemUtilities.Write(bufferStream, (int)SerializeId.CreationTime, this.CreationTime.ToUniversalTime().ToString("yyyy-MM-ddTHH:mm:ssZ", System.Globalization.DateTimeFormatInfo.InvariantInfo));
             }
 
             // Key
@@ -104,7 +104,7 @@ namespace Library.Net.Outopos
             {
                 using (var stream = this.Key.Export(bufferManager))
                 {
-                    ItemUtilities.Write(bufferStream, (byte)SerializeId.Key, stream);
+                    ItemUtilities.Write(bufferStream, (int)SerializeId.Key, stream);
                 }
             }
 
@@ -113,7 +113,7 @@ namespace Library.Net.Outopos
             {
                 using (var stream = this.Cash.Export(bufferManager))
                 {
-                    ItemUtilities.Write(bufferStream, (byte)SerializeId.Cash, stream);
+                    ItemUtilities.Write(bufferStream, (int)SerializeId.Cash, stream);
                 }
             }
             // Certificate
@@ -121,7 +121,7 @@ namespace Library.Net.Outopos
             {
                 using (var stream = this.Certificate.Export(bufferManager))
                 {
-                    ItemUtilities.Write(bufferStream, (byte)SerializeId.Certificate, stream);
+                    ItemUtilities.Write(bufferStream, (int)SerializeId.Certificate, stream);
                 }
             }
