+++ Library.Net.Outopos/Cache/Metadata/Key.cs
@@ -66,13 +66,13 @@ namespace Library.Net.Outopos
             // Hash
             if (this.Hash != null)
             {
-                ItemUtilities.Write(bufferStream, (byte)SerializeId.Hash, this.Hash);
+                ItemUtilities.Write(bufferStream, (int)SerializeId.Hash, this.Hash);
             }
 
             // HashAlgorithm
             if (this.HashAlgorithm != 0)
             {
-                ItemUtilities.Write(bufferStream, (byte)SerializeId.HashAlgorithm, this.HashAlgorithm.ToString());
+                ItemUtilities.Write(bufferStream, (int)SerializeId.HashAlgorithm, this.HashAlgorithm.ToString());
             }
 
             bufferStream.Seek(0, SeekOrigin.Begin);
